+++ Libraries/ResultCollectionAccess/include/ResultCollectionExtender.h
@@ -29,7 +29,7 @@ public:
 	void StoreCampaignChanges();
 	void FlushQuantityContainer();
 
-	void AddQuantityContainer(uint32_t seriesIndex, std::list<std::string>& parameterAbbreviations, std::list<ot::Variable>&& parameterValues, uint32_t quantityIndex, const ot::Variable& quantityValue);
+	void AddQuantityContainer(uint64_t seriesIndex, std::list<std::string>& parameterAbbreviations, std::list<ot::Variable>&& parameterValues, uint64_t quantityIndex, const ot::Variable& quantityValue);
 
 private:
 	friend class FixtureResultCollectionExtender;
